divide the app in more components, fix some small css details and add date validation in the forms
