+++ src/components/NewTrip.js
@@ -1,8 +1,9 @@
 import "./styles/NewTrip.css";
 import { useHistory } from "react-router";
-import { Link } from "react-router-dom";
+
 import { useState } from "react";
 import { addTripToLocalStorage } from "../services/myTripsStorage";
+import NewTripForm from "./NewTripForm";
 
 export default function NewTrip() {
   const [inputDestinationName, setInputDestinationName] = useState("");
@@ -48,166 +49,26 @@ export default function NewTrip() {
       <div className="newTripHeader">
         <h2>NEW TRIP</h2>
       </div>
-
-      <form className="NewTripForm" onSubmit={handleOnSubmit}>
-        <div className="formHeader">
-          <label className="tripNameForm" htmlFor="tripName">
-            Destination Name
-            <input
-              onChange={(e) => {
-                setInputDestinationName(e.target.value);
-              }}
-              value={inputDestinationName}
-              id="tripName"
-              type="text"
-              placeholder="Type your destination"
-              maxLength="15"
-              required
-            ></input>
-          </label>
-        </div>
-        <div className="formTravelDates">
-          <label className="tripStartForm" htmlFor="tripStart">
-            START
-            <input
-              onChange={(e) => {
-                setInputTripStart(e.target.value);
-              }}
-              id="tripStart"
-              type="date"
-              required
-            ></input>
-          </label>
-
-          <label className="tripEndForm" htmlFor="tripEnd">
-            END
-            <input
-              onChange={(e) => {
-                setInputTripEnd(e.target.value);
-              }}
-              id="tripEnd"
-              type="date"
-              required
-            ></input>
-          </label>
-        </div>
-        <div className="formTransport">
-          <label className="tripTransportForm" htmlFor="tripTransportType">
-            TRANSPORT
-            <select
-              onChange={(e) => {
-                setInputTransportType(e.target.value);
-              }}
-              id="tripTransportType"
-            >
-              <option value="none">Transport Type</option>
-              <option value="bus">Bus</option>
-              <option value="car">Car</option>
-              <option value="plane">Plane</option>
-              <option value="train">Train</option>
-              <option value="other">Other</option>
-            </select>
-          </label>
-
-          <div className="departureArrival">
-            <label className="inputTranportTimeForm" htmlFor="tripDeparture">
-              DEPARTURE TIME
-              <input
-                onChange={(e) => {
-                  setInputTripDeparture(e.target.value);
-                }}
-                id="tripDeparture"
-                type="time"
-              ></input>
-            </label>
-
-            <label className="inputTranportTimeForm" htmlFor="tripArrival">
-              ARRIVAL TIME
-              <input
-                onChange={(e) => {
-                  setInputTripArrival(e.target.value);
-                }}
-                id="tripArrival"
-                type="time"
-              ></input>
-            </label>
-          </div>
-        </div>
-        <div className="formAccommodation">
-          <label
-            className="inputAccommodationFormName"
-            htmlFor="tripAccommodation"
-          >
-            ACCOMMODATION
-            <input
-              onChange={(e) => {
-                setInputTripAccommodation(e.target.value);
-              }}
-              id="tripAccommodation"
-              placeholder="Type your accommodation"
-              type="text"
-            ></input>
-          </label>
-          <div className="checkinCheckout">
-            <label className="inputAccommodationForm" htmlFor="tripCheckin">
-              CHECK-IN
-              <input
-                onChange={(e) => {
-                  setInputCheckinDate(e.target.value);
-                }}
-                id="tripCheckinDate"
-                type="date"
-              ></input>
-              <input
-                className="accommodationTime"
-                onChange={(e) => {
-                  setInputCheckinTime(e.target.value);
-                }}
-                id="tripCheckinTime"
-                type="time"
-              ></input>
-            </label>
-
-            <label className="inputAccommodationForm" htmlFor="tripCheckout">
-              CHECK-OUT
-              <input
-                onChange={(e) => {
-                  setInputCheckoutDate(e.target.value);
-                }}
-                id="tripCheckoutDate"
-                type="date"
-              ></input>
-              <input
-                className="accommodationTime"
-                onChange={(e) => {
-                  setInputCheckoutTime(e.target.value);
-                }}
-                id="tripCheckoutTime"
-                type="time"
-              ></input>
-            </label>
-          </div>
-        </div>
-        <div className="formSightseeing">
-          <label className="inputSightseeingForm" htmlFor="sightseeingList">
-            <textarea
-              onChange={(e) => {
-                setInputSightseeing(e.target.value);
-              }}
-              id="sightseeingList"
-              placeholder="Type your travel notes here..."
-            ></textarea>
-          </label>
-        </div>
-        <div className="newTripFormFooter">
-          <button type="submit" className="createNewTripButton">
-            Create
-          </button>
-          <Link className="cancelNewTrip" to="/myTrips">
-            <button className="cancelNewTripButton">Cancel</button>
-          </Link>
-        </div>
-      </form>
+      <NewTripForm
+        handleOnSubmit={handleOnSubmit}
+        inputDestinationName={inputDestinationName}
+        setInputDestinationName={setInputDestinationName}
+        inputTripStart={inputTripStart}
+        setInputTripStart={setInputTripStart}
+        inputTripEnd={inputTripEnd}
+        setInputTripEnd={setInputTripEnd}
+        setInputTransportType={setInputTransportType}
+        setInputTripDeparture={setInputTripDeparture}
+        setInputTripArrival={setInputTripArrival}
+        setInputTripAccommodation={setInputTripAccommodation}
+        inputCheckinDate={inputCheckinDate}
+        setInputCheckinDate={setInputCheckinDate}
+        setInputCheckinTime={setInputCheckinTime}
+        inputCheckoutDate={inputCheckoutDate}
+        setInputCheckoutDate={setInputCheckoutDate}
+        setInputCheckoutTime={setInputCheckoutTime}
+        setInputSightseeing={setInputSightseeing}
+      />
     </div>
   );
 }
