+++ src/components/styles/MyTrips.css
@@ -171,11 +171,6 @@
   background-color: rgb(0, 78, 0);
 }
 
-.transportTypeIcon {
-  text-decoration: none;
-  color: black;
-}
-
 .tripDates {
   position: absolute;
   right: 33px;
@@ -188,7 +183,7 @@ hr {
   transform: rotate(-90deg);
   width: 82px;
   position: absolute;
-  right: 55px;
+  top: 47px;
 }
 
 .barcode {
@@ -236,7 +231,14 @@ hr {
     text-decoration: none;
     font-size: 10px;
     color: black;
+    position: relative;
   }
+
+  .tripCardInfo {
+    display: flex;
+    align-items: center;
+  }
+
   .tripDates {
     right: 32px;
     top: 8px;
@@ -253,7 +255,7 @@ hr {
   hr {
     right: 40px;
   }
-  .transportTypeIcon {
+  .tripCardInfo i {
     display: none;
   }
 }
@@ -284,7 +286,7 @@ hr {
     font-size: 16px;
     margin-left: 10px;
   }
-  .transportTypeIcon {
+  .tripCardInfo i {
     display: inline;
   }
 }
